+++ flower-shop-frontend/src/pages/ErrorPage.tsx
@@ -1,5 +1,9 @@
 import * as React from 'react';
 import { isRouteErrorResponse, useRouteError } from 'react-router-dom';
+import Header from '../components/Header';
+import Footer from '../components/Footer';
+import Grid from '@mui/material/Grid';
+import Typography from '@mui/material/Typography';
 
 export default function ErrorPage() {
   const error = useRouteError();
@@ -8,21 +12,53 @@ export default function ErrorPage() {
   if (isRouteErrorResponse(error)) {
     return (
       <div id="error-page">
-        <h1>Oops!</h1>
-        <p>Sorry, an error has occurred.</p>
-        <p>
-          <i>
-            {error.status} {error.statusText}
-          </i>
-        </p>
+        <Header />
+        <Grid
+          container
+          display="flex"
+          direction="column"
+          alignItems="center"
+          justifyContent="center"
+          spacing={2}
+          width="100%"
+          minHeight="calc(100vh - 203px - 70px)"
+          sx={{ pr: 4, pl: 4, zIndex: 1, maxWidth: '100%' }}
+        >
+          <Typography variant="h2" marginTop="10%">
+            Oops!
+          </Typography>
+          <Typography variant="h6">Sorry, an unknown error has occurred.</Typography>
+          <p>
+            <i>
+              {error.status} {error.statusText}
+            </i>
+          </p>
+        </Grid>
+        <Footer />
       </div>
     );
   }
 
   return (
     <div id="error-page">
-      <h1>Oops!</h1>
-      <p>Sorry, an unknown error has occurred.</p>
+      <Header />
+      <Grid
+        container
+        display="flex"
+        direction="column"
+        alignItems="center"
+        justifyContent="center"
+        spacing={2}
+        width="100%"
+        minHeight="calc(100vh - 203px - 70px)"
+        sx={{ pr: 4, pl: 4, zIndex: 1, maxWidth: '100%' }}
+      >
+        <Typography variant="h2" marginTop="10%">
+          Oops!
+        </Typography>
+        <Typography variant="h6">Sorry, an unknown error has occurred.</Typography>
+      </Grid>
+      <Footer />
     </div>
   );
 }
